# memdb: merge nodeData and kvData to speedup access